chore: move to using @imgix/js-core
