+++ src/common/imgix-js-core-wrapper.ts
@@ -1,5 +1,5 @@
+import ImgixClient from '@imgix/js-core';
 import * as E from 'fp-ts/lib/Either';
-import ImgixClient from 'imgix-core-js';
 
 export const createImgixClient = (
   options: ConstructorParameters<typeof ImgixClient>[0],
