+++ test/unit/source-url/source-url.test.ts
@@ -116,7 +116,7 @@ describe('createResolvers', () => {
           field: 'fluid',
         });
 
-        // Don't need to do too much work here since imgix-core-js handles everything under the hood
+        // Don't need to do too much work here since @imgix/js-core handles everything under the hood
         expect(fluidFieldResult.srcSet).toMatch(/\d*w,/);
         expect(fluidFieldResult.srcSetWebp).toMatch(/\d*w,/);
       });
@@ -196,7 +196,7 @@ describe('createResolvers', () => {
           field: 'fluid',
         });
 
-        // Don't need to do too much work here since imgix-core-js handles everything under the hood
+        // Don't need to do too much work here since @imgix/js-core handles everything under the hood
         expect(fluidFieldResult.src).not.toMatch('ar=');
         expect(fluidFieldResult.srcWebp).not.toMatch('ar=');
         expect(fluidFieldResult.srcSet).not.toMatch('ar=');
@@ -208,7 +208,7 @@ describe('createResolvers', () => {
           fieldParams: { maxWidth: 500 },
         });
 
-        // Don't need to do too much work here since imgix-core-js handles everything under the hood
+        // Don't need to do too much work here since @imgix/js-core handles everything under the hood
         expect(fluidFieldResult.src).toMatch('w=500');
         expect(fluidFieldResult.srcWebp).toMatch('w=500');
       });
@@ -229,7 +229,7 @@ describe('createResolvers', () => {
           );
         };
 
-        // Don't need to do too much work here since imgix-core-js handles everything under the hood
+        // Don't need to do too much work here since @imgix/js-core handles everything under the hood
         expectSrcsetToNotHaveWidthsGT(500)(fluidFieldResult.srcSet);
         expectSrcsetToNotHaveWidthsGT(500)(fluidFieldResult.srcSetWebp);
       });
